refactor(packet): use jackpal/gateway for route discovery, lazy-init pcap
Replace DNS-probe + packet-capture based gateway detection with jackpal/gateway
OS route table lookup. ARP MAC resolution is now skipped when NeedsPcap() is
false. FakeCount narrowed from int to uint8. Constructor signatures cleaned up
across sniffer, writer, and proxy types.

+++ internal/config/cli_test.go
@@ -42,7 +42,7 @@ func TestCreateCommand_Flags(t *testing.T) {
 				assert.Equal(t, HTTPSSplitModeSNI, cfg.Runtime.HTTPS.SplitMode)
 				assert.Equal(t, uint8(35), cfg.Runtime.HTTPS.ChunkSize)
 				assert.False(t, cfg.Runtime.HTTPS.Skip)
-				assert.Equal(t, 0, cfg.Runtime.UDP.FakeCount)
+				assert.Equal(t, uint8(0), cfg.Runtime.UDP.FakeCount)
 				assert.Equal(t, 64, len(cfg.Runtime.UDP.FakePacket))
 			},
 		},
@@ -106,7 +106,7 @@ func TestCreateCommand_Flags(t *testing.T) {
 				assert.True(t, cfg.Runtime.HTTPS.Skip)
 
 				// UDP
-				assert.Equal(t, 5, cfg.Runtime.UDP.FakeCount)
+				assert.Equal(t, uint8(5), cfg.Runtime.UDP.FakeCount)
 				assert.Equal(t, []byte{0x01, 0x02}, cfg.Runtime.UDP.FakePacket)
 			},
 		},
@@ -305,7 +305,7 @@ func TestCreateCommand_OverrideTOML(t *testing.T) {
 	assert.False(t, capturedCfg.Runtime.HTTPS.Skip)
 
 	// UDP
-	assert.Equal(t, 20, capturedCfg.Runtime.UDP.FakeCount)
+	assert.Equal(t, uint8(20), capturedCfg.Runtime.UDP.FakeCount)
 	assert.Equal(t, []byte{0xcc, 0xdd}, capturedCfg.Runtime.UDP.FakePacket)
 	assert.Equal(t, []byte{0xcc, 0xdd}, capturedCfg.Runtime.UDP.FakePacket)
